src/nyc_trees/requirements/development.txt
-r base.txt

coverage==3.6
django-debug-toolbar==1.2
sqlparse==0.1
Sphinx==1.2b1
ipdb==0.8
